Created dialogue functionality for the Goddess
Several text documents were added that contain lines of dialogue.

For now, it reads from a test file.

Instead of using a timer to make the dialogue box go away, the player presses SPACE to go to the next line.

This test dialogue is accessible by collecting the TestCollectible object and then talking to the Goddess.

The Dialogue Box was also updated with larger text size and wrapping so that the lines of dialogue are easier to read and stay inside the label.