Dialogue implementation and collectables
Began working on implementing the rest of the dialogue for the Goddess. It does print the correct group of dialogue based on how many collectables the player has.

The the three collectable objects have also been created.

The dialogue function is not complete just yet, as it needs to tell the player which collectables the player still needs to progress.