Finishing touches
The two levels now have a collectable at the end.

Collecting items now sends the player back to the beginning of the hub level.

Made the AI in level 2 move a little slower.

The levels now check if the collectable is present. This ensures the player can't  reenter a level if the level's collectable has already been collected.